+++ include/swift/Basic/RelativePointer.h
@@ -132,7 +132,9 @@
 #ifndef SWIFT_BASIC_RELATIVEPOINTER_H
 #define SWIFT_BASIC_RELATIVEPOINTER_H
 
+#include <cassert>
 #include <cstdint>
+#include <type_traits>
 
 namespace swift {
